+++ packages/yourloops/test/unit/lib/error/error-api.test.tsx
@@ -1,6 +1,5 @@
 /**
- * Copyright (c) 2021, Diabeloop
- * Main App file
+ * Copyright (c) 2022, Diabeloop
  *
  * All rights reserved.
  *
@@ -26,44 +25,27 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-// Polyfills for compatibility with older browsers:
-import 'core-js/stable'
+import HttpService from '../../../../services/http'
+import ErrorApi, { ErrorPayload } from '../../../../lib/error/error-api'
 
-import React from 'react'
-import ReactDOM from 'react-dom'
+describe('ErrorApi', () => {
+  describe('sendError', () => {
+    it('should send correct payload to correct url', async () => {
+      const payload: ErrorPayload = {
+        browserName: 'fakeBrowserName',
+        browserVersion: 'fakeBrowserVersion',
+        date: 'fakeDate',
+        err: 'fakeErrorMessage',
+        errorId: 'fakeErrorId',
+        path: '/fake/path'
+      }
+      jest.spyOn(HttpService, 'post').mockResolvedValueOnce(null)
 
-import config from '../lib/config'
-import { init as i18nInit } from '../lib/language'
-import initCookiesConcentListener from '../lib/cookies-manager'
-import initDayJS from '../lib/dayjs'
-import initAxios from '../lib/axios'
-import { initTheme } from '../components/theme'
-
-import Yourloops from './app'
-import OnError from './error'
-import { BrowserRouter } from 'react-router-dom'
-
-i18nInit().then(() => {
-  window.onerror = (event, source, lineno, colno, error) => {
-    if (source && !source.endsWith('.js')) {
-      return true
-    }
-    console.error(event, source, lineno, colno, error)
-    ReactDOM.render(<BrowserRouter><OnError event={event} source={source} lineno={lineno} colno={colno} error={error} /></BrowserRouter>, document.body)
-    return false
-  }
-
-  let div = document.getElementById('app')
-  if (div === null) {
-    div = document.createElement('div')
-    div.id = 'app'
-    document.body.appendChild(div)
-  }
-
-  initDayJS()
-  initCookiesConcentListener()
-  initAxios()
-  initTheme()
-
-  ReactDOM.render(config.DEV ? <React.StrictMode><Yourloops /></React.StrictMode> : <Yourloops />, div)
+      await ErrorApi.sendError(payload)
+      expect(HttpService.post).toHaveBeenCalledWith({
+        url: '/bff/v1/errors',
+        payload
+      })
+    })
+  })
 })